Introducing new base form for app analytics.
